refactor: 패키지 구조 변경과 Validator 로직분리, 상수 추출
- 패키지 구조를 domain 형식으로 변경
- 중첩되어 있던 상수를 클래스 화하여 중복제거
- 검증 로직을 분리
