+++ src/main/java/vendingmachine/model/coin/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine.model.manage;
+package vendingmachine.model.coin;
 
 import java.util.Arrays;
 import java.util.LinkedHashMap;
