+++ src/main/java/vendingmachine/model/coin/MachineMoney.java
@@ -1,4 +1,4 @@
-package vendingmachine.model.manage;
+package vendingmachine.model.coin;
 
 import java.util.Map;
