+++ src/main/java/vendingmachine/model/drink/DrinkValidator.java
@@ -1,4 +1,4 @@
-package vendingmachine.model.order;
+package vendingmachine.model.drink;
 
 import vendingmachine.util.Constant;
 import vendingmachine.util.ErrorMessage;
