+++ src/main/java/vendingmachine/model/drink/Drinks.java
@@ -1,4 +1,4 @@
-package vendingmachine.model.order;
+package vendingmachine.model.drink;
 
 import java.util.List;
