+++ src/main/java/vendingmachine/util/Constant.java
@@ -10,6 +10,7 @@ public class Constant {
     public static final int NAME_INDEX = 0;
     public static final int PRICE_INDEX = 1;
     public static final int QUANTITY_INDEX = 2;
+    public static final int CHECK_ARR_HAS_ALL_CONTENT = 3;
 
     public static final String CHANGE_BRACKET_TO_GAP = "";
     public static final String OPEN_SQUARE_BRACKET = "[";
